[SC-80] refactor join popup into a component
